risk-service/requirements.txt
# 风控合规服务 - 依赖 (risk-service)
# Python 3.11 + FastAPI + Elasticsearch 8.x

fastapi>=0.104.0,<0.115.0
uvicorn[standard]>=0.24.0,<0.32.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Elasticsearch 8.x
elasticsearch[async]>=8.11.0,<9.0.0

# 配置与工具
pyyaml>=6.0.0,<7.0.0

# 可选: 测试
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<0.24.0
pytest-cov>=4.1.0,<6.0.0
httpx>=0.25.0,<0.28.0
